Extract modals for PHP and server
